test: add image OCR regression
