+++ tests/MarkItDownNet.Tests/OcrTestHelpers.cs
@@ -1,4 +1,5 @@
 using System;
+using TesseractOCR.InteropDotNet;
 
 namespace MarkItDownNet.Tests;
 
@@ -7,5 +8,6 @@ internal static class OcrTestHelpers
     public static void EnsureOcrLibraries()
     {
         Environment.SetEnvironmentVariable("LD_LIBRARY_PATH", "/usr/lib/x86_64-linux-gnu");
+        LibraryLoader.Instance.CustomSearchPath = "/usr/lib/x86_64-linux-gnu";
     }
 }
